Moves things around, decompose into modules (WIP)
