+++ coursera/downloaders.py
@@ -12,13 +12,17 @@ from __future__ import print_function
 import logging
 import math
 import os
-import requests
 import subprocess
 import sys
 import time
 
+import requests
+
 from six import iteritems
 
+#
+# Below are file downloaders, they are wrappers for external downloaders.
+#
 
 class Downloader(object):
     """
